add day 2 of aoc 2022
This implementation can be greatly optimized. I am utilizing nested for loops and look ups to check if a specific move meets a needed criteria. The fastest way to solve this would likely be to pre-compile all possible move combinations as well as the related player score for that move combination. My current implementation, however, could be easier to expand/modify logic and was the least time-consuming way I could think of writing this. If I refactor, I will likely create pre-compiled outcomes, or perhaps do some math with character ordinal values in order to reduce iteration and not be dependent on as many hard-coded mappings.